Update search_candidates.R
There's a weird thing going on with a NULL value for this particular committee's treasurer_name. .default arguments for the map_chr function aren't working, need to spend some time with Advanced R to up my knowledge of what exactly is going on. For now, I've replaced the map_chr function with map to pull a list and as.character to coerce it. Result is for this one committee, the string "NULL" is returned for the treasurer name. Because this isn't the most pressing feature and it should be relatively clear for anyone viewing the string, I'm implementing this so that search_candidates doesn't break down for now.